move components to a separate files - move types to a separate files
