+++ src/components/CommentInfo/CommentInfo.tsx
@@ -1,5 +1,34 @@
-import React from 'react';
+import { Comment } from '../../types/Comment';
 
-export const CommentInfo: React.FC = () => (
-  <>Put the comment here</>
-);
+interface CommentInfoProps {
+  postComment: Comment;
+}
+
+export const CommentInfo: React.FC<CommentInfoProps> = ({ postComment }) => {
+  const {
+    name,
+    email,
+    body,
+  } = postComment;
+
+  return (
+    <div className="CommentInfo">
+      <div className="CommentInfo__title">
+        <strong className="CommentInfo__name">{name}</strong>
+
+        {' by '}
+
+        <a
+          className="CommentInfo__email"
+          href={`mailto:${email}`}
+        >
+          Telly_Lynch@karl.co.uk
+        </a>
+      </div>
+
+      <div className="CommentInfo__body">
+        {body}
+      </div>
+    </div>
+  );
+};
