+++ src/components/CommentList/CommentList.tsx
@@ -1,5 +1,25 @@
-import React from 'react';
+import { Comment } from '../../types/Comment';
+import { CommentInfo } from '../CommentInfo';
 
-export const CommentList: React.FC = () => (
-  <>Put the list here</>
-);
+interface CommentListProps {
+  commentList: Comment[];
+}
+
+export const CommentList: React.FC<CommentListProps> = ({ commentList }) => {
+  const isHaveComments = commentList.length > 0;
+
+  return (isHaveComments
+    ? (
+      <div className="CommentList">
+        {commentList.map(comment => (
+          <CommentInfo postComment={comment} key={comment.id} />
+        ))}
+      </div>
+    ) : (
+      <>
+        <hr />
+        <b data-cy="NoCommentsMessage">No comments yet</b>
+      </>
+    )
+  );
+};
