+++ src/types/FullPostData.ts
@@ -1,0 +1,8 @@
+import { Comment } from './Comment';
+import { Post } from './Post';
+import { User } from './User';
+
+export interface FullPostData extends Post {
+  user?: User;
+  comments: Comment[];
+}
